Migrate to EB v3.x CLI and Frankfurt region

+++ res/aws_release.sh
@@ -4,16 +4,22 @@
 cd `dirname $0`
 cd ..
 
+if [ "$#" -ne 3 ] || [ "$1" != "prod" ] && [ "$1" != "test" ] ; then
+    echo 'Usage: ./res/aws_release prod|test from-branch 1.0.0'
+    exit 0
+fi
+
 # switch branch
-git checkout aws-dist
-git merge master --no-edit
+git checkout $2
+git branch release/$1
+git checkout release/$1
+git merge $2 --no-edit
 
 # build and test
 rm -rf node_modules/
 npm cache clear
 npm install
-grunt release-$1 --release=$2
-grunt test
+grunt release-$1 --release=$3
 
 # install only production dependencies
 rm -rf node_modules/
@@ -26,10 +32,10 @@ find node_modules/ -name ".gitignore" -exec rm -rf {} \;
 sed -i "" '/dist/d' .gitignore
 sed -i "" '/node_modules/d' .gitignore
 git add .gitignore node_modules/ dist/
-git commit -m "Update aws-dist"
+git commit -m "Update release"
 
 # push to aws
-git aws.push
+eb deploy
 
-# switch back to master branch
-git checkout master
+# switch back to $2 branch
+git checkout $2
